app/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=1.10.0
underthesea>=1.3.5
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1
google-generativeai
python-dotenv
openpyxl